+++ Run_Pipeline/Main_Pipeline.py
@@ -4,13 +4,13 @@ from langchain.agents import create_tool_calling_agent, AgentExecutor
 from langchain_community.chat_message_histories import ChatMessageHistory
 from langchain_core.runnables.history import RunnableWithMessageHistory
 from langchain_teddynote.messages import AgentCallbacks, AgentStreamParser
-from Agnet_Tools import *
+from Agent_Tool import *
 from Document_Loader import DocumentLoader
 from Document_Splitter import DocumentSplitter
 from Embedding_DB import EmbeddingDB
 from Env_Loader import EnvLoader
 from LLM_Factory import LLMFactory
-from Retiever_Builder import RetrieverBuilder
+from Retriever_Builder import RetrieverBuilder
 
 
 def main(return_chain_only: bool = False):
@@ -49,10 +49,11 @@ def main(return_chain_only: bool = False):
         overlap_size = 50
         persist_dir = BASE_DIR / os.getenv("DATA_PATH") / f"{kind}_{chunk_size}"
         device = "mps"
-        retriever_mode = 1
-        k = 3
+        retriever_mode = 3
+        k = 10
         engine_num = 1
         backend_num = 1
+
     # ③ 문서 로드
     loader = DocumentLoader(file_path, kind)  # 일반 파일들을 Document형태로 변환해서 다 불러옴
     docs = loader._load_json_files()
@@ -70,14 +71,18 @@ def main(return_chain_only: bool = False):
     llm = LLMFactory.load_llm(engine=engine_num, backend=backend_num, device=device)  # LLM 로드
 
     # ⑥ 리트리버 빌드
-    retriever_builder = RetrieverBuilder(db=db, docs=chunks, k=k, mode=retriever_mode)  # 리트리버 빌더 생성
+    retriever_builder = RetrieverBuilder(
+        db=db,
+        docs=chunks,
+        k=k,
+        mode=retriever_mode
+    )
     retriever = retriever_builder.build()
 
     tools = AgentTools.get_tools(retriever=retriever, llm=llm)  # 도구 목록 생성 (retriever 포함)
 
     agent = create_tool_calling_agent(llm, tools, AgentTools.PROMPT)
     agent_executor = AgentExecutor(agent=agent, tools=tools, verbose=False)
-
 
     # 도구 호출 시 실행되는 콜백 함수입니다.
     def tool_callback(tool) -> None:
@@ -108,7 +113,6 @@ def main(return_chain_only: bool = False):
 
     # AgentStreamParser 객체를 생성하여 에이전트의 실행 과정을 파싱합니다.
     agent_stream_parser = AgentStreamParser(agent_callbacks)
-
 
     store = {}
 
@@ -144,6 +148,7 @@ def main(return_chain_only: bool = False):
         for step in result:
             agent_stream_parser.process_agent_steps(step)
         print()  # 다음 질문을 위해 줄바꿈
+
 
 if __name__ == "__main__":
     main()
